Move Profile Model to Profile Folder
