+++ app/Http/Controllers/Api/V1/AuthController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Models\Profile;
+use App\Models\Profile\Profile;
 use App\Models\User\User;
 use App\Notifications\Activated;
 use App\Notifications\Activation;
