+++ app/Http/Controllers/Api/V1/SocialAuthController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Models\Profile;
+use App\Models\Profile\Profile;
 use App\Models\User\User;
 use JWTAuth;
 use Socialite;
