+++ app/Models/Profile/Profile.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Profile;
 
 use App\Models\Profile\Traits\Relationship\ProfileRelationship;
 use Eloquent;
